requirements.txt
apache-airflow==2.9.3
pendulum==3.0.0

pytest==8.3.2

python-dotenv==1.0.1
pydantic==2.8.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
SQLAlchemy==1.4.53

duckdb==1.0.0
